qemux86: configure rauc
This adds a proper system.conf for the setup, describing the EFI boot
slot and rescue system for reference, and defining the redundant root
file system slots to be sda5 (rootfs.0) and sda6 (rootfs.1).

The common data partition (sda7) mounted to /data will be used for
storing the rauc status file.