qemux86: add boot-image
While creating an image for the boot partition is straight-forward when
using wic during disk image creation, it is not when we require a separate
file system image in DEPLOYDIR (for updating the boot partition).

As wic does not provide intermediate results as file system images, we
need to create our image manually for now.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>