qemux86: move grub environment to extra partition
This is a preparation for enabling atomic bootloader updates.

Having the grub environment in the same partition as the bootloader
would drop the environment each time we update the bootloader.

Instead, we place it on a single, separate partition, let grub point to
this and mount this to /efienv in Linux userspace to allow
manipulating it.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>